Move simplification tree to fuzzer module
