+++ fuzzer/api/src/commonTest/kotlin/com/github/jcornaz/kwik/fuzzer/api/simplifier/tree/TreeAssertion.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.kwik.generator.api.simplification
+package com.github.jcornaz.kwik.fuzzer.api.simplifier.tree
 
 import com.github.jcornaz.kwik.generator.api.ExperimentalKwikGeneratorApi
 import kotlin.test.assertEquals
